+++ launchkey/clients/base.py
@@ -8,7 +8,9 @@ from formencode import Invalid
 from launchkey.entities.shared import PublicKey
 
 from ..utils.shared import iso_format
-from ..entities.service import Service, ServiceSecurityPolicy
+from ..entities.service import Service, ServiceSecurityPolicy, \
+    ConditionalGeoFencePolicy, GeoCircleFence, TerritoryFence, \
+    MethodAmountPolicy, FactorsPolicy
 from ..entities.validation import ServiceValidator, PublicKeyValidator, \
     ServiceSecurityPolicyValidator
 
@@ -20,7 +22,7 @@ from ..exceptions import InvalidEntityID, LaunchKeyAPIException, \
     InvalidPublicKey, PublicKeyDoesNotExist, LastRemainingKey, \
     LastRemainingSDKKey, InvalidSDKKey, DirectoryNameInUse, \
     AuthorizationInProgress, Conflict, AuthorizationResponseExists, \
-    AuthorizationRequestCanceled
+    AuthorizationRequestCanceled, NestedPolicyTypeError, UnknownPolicyException
 from ..transports.base import APIResponse
 
 ERROR_CODE_MAP = {
@@ -388,9 +390,93 @@ class ServiceManagingBaseClient(BaseClient):
         policy_data = self._validate_response(response.data,
                                               ServiceSecurityPolicyValidator)
 
-        policy = ServiceSecurityPolicy()
-        policy.set_policy(policy_data)
+        if policy_data["type"] == "LEGACY":
+            policy = ServiceSecurityPolicy()
+            policy.set_policy(policy_data)
+        else:
+            fences = list()
+            for fence in policy_data["fences"]:
+                if "type" in fence:
+                    if fence["type"] == "GEO_CIRCLE":
+                        fences.append(
+                            GeoCircleFence(fence["latitude"],
+                                           fence["longitude"],
+                                           fence["radius"]
+                                           )
+                        )
+                    elif fence["type"] == "TERRITORY":
+                        fences.append(
+                            TerritoryFence(fence["country"],
+                                           fence["administrative_area"],
+                                           fence["postal_code"]
+                                           )
+                        )
+                    elif fence["type"] == "GEO_FENCE":
+                        raise NotImplementedError
+
+            if policy_data["type"] == "COND_GEO":
+                inside = self.__process_nested_service_policy(
+                    policy_data["inside"]
+                )
+                outside = self.__process_nested_service_policy(
+                    policy_data["outside"]
+                )
+
+                policy = ConditionalGeoFencePolicy(
+                    inside,
+                    outside,
+                    policy_data["deny_rooted_jailbroken"],
+                    policy_data["deny_emulator_simulator"],
+                    fences
+                )
+            elif policy_data["type"] == "METHOD_AMOUNT":
+                policy = MethodAmountPolicy(
+                    deny_rooted_jailbroken=policy_data[
+                        "deny_rooted_jailbroken"],
+                    deny_emulator_simulator=policy_data[
+                        "deny_emulator_simulator"],
+                    fences=fences,
+                    amount=policy_data["amount"]
+                )
+            elif policy_data["type"] == "FACTORS":
+                policy = FactorsPolicy(
+                    deny_rooted_jailbroken=policy_data[
+                        "deny_rooted_jailbroken"],
+                    deny_emulator_simulator=policy_data[
+                        "deny_emulator_simulator"],
+                    fences=fences,
+                    factors=policy_data["factors"]
+                )
+            else:
+                raise UnknownPolicyException(
+                    "The Policy {0} was not a known Policy type".format(
+                        policy_data["type"])
+                )
+
         return policy
+
+    @staticmethod
+    def __process_nested_service_policy(policy):
+        if policy["type"] == "METHOD_AMOUNT":
+            new_policy = MethodAmountPolicy(
+                amount=policy["amount"],
+                deny_rooted_jailbroken=None,
+                deny_emulator_simulator=None,
+                fences=policy["fences"]
+            )
+        elif policy["type"] == "FACTORS":
+            new_policy = FactorsPolicy(
+                factors=policy["factors"],
+                deny_rooted_jailbroken=None,
+                deny_emulator_simulator=None,
+                fences=policy["fences"]
+            )
+        else:
+            raise NestedPolicyTypeError(
+                "Valid nested Policy types for ConditionalGeofence Policies "
+                "are: [\"METHOD_AMOUNT\", \"FACTORS\"]"
+            )
+        return new_policy
 
     @api_call
     def set_service_policy(self, service_id, policy):
@@ -407,7 +493,7 @@ class ServiceManagingBaseClient(BaseClient):
         self._transport.put(
             "{}/policy".format(self.__service_base_path[0:-1]),
             self._subject, service_id=str(service_id),
-            policy=policy.get_policy())
+            policy=policy.to_json())
 
     @api_call
     def remove_service_policy(self, service_id):